Add unit test for Bookmark.create
Bookmark also has a method for determining the appropriate
connection to use, which is now private.

The feature spec for adding_bookmarks had a redundant test for
elements on the page; this test has been removed. Note there was
some example code for using Nokogiri which we may want to revisit
in future!